+++ packages/client/src/Call.ts
@@ -107,6 +107,7 @@ import {
 import { BehaviorSubject, Subject, takeWhile } from 'rxjs';
 import { ReconnectDetails } from './gen/video/sfu/event/events';
 import {
+  ClientCapability,
   ClientDetails,
   Codec,
   PublishOption,
@@ -270,6 +271,13 @@ export class Call {
 
   private readonly streamClientBasePath: string;
   private streamClientEventHandlers = new Map<Function, () => void>();
+
+  /**
+   * A list of capabilities that the client supports and are enabled.
+   */
+  private clientCapabilities = new Set<ClientCapability>([
+    ClientCapability.SUBSCRIBER_VIDEO_PAUSE,
+  ]);
 
   /**
    * Constructs a new `Call` instance.
@@ -987,6 +995,7 @@ export class Call {
             reconnectDetails,
             preferredPublishOptions,
             preferredSubscribeOptions,
+            capabilities: Array.from(this.clientCapabilities),
           });
 
         this.currentPublishOptions = publishOptions;
@@ -2745,5 +2754,23 @@ export class Call {
    */
   setDisconnectionTimeout = (timeoutSeconds: number) => {
     this.disconnectionTimeoutSeconds = timeoutSeconds;
+  };
+
+  /**
+   * Enables the provided client capabilities.
+   */
+  enableClientCapabilities = (...capabilities: ClientCapability[]) => {
+    for (const capability of capabilities) {
+      this.clientCapabilities.add(capability);
+    }
+  };
+
+  /**
+   * Disables the provided client capabilities.
+   */
+  disableClientCapabilities = (...capabilities: ClientCapability[]) => {
+    for (const capability of capabilities) {
+      this.clientCapabilities.delete(capability);
+    }
   };
 }