+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.7.8](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-1.7.7...@stream-io/video-react-bindings-1.7.8) (2025-07-17)
+
+### Dependency Updates
+
+- `@stream-io/video-client` updated to version `1.26.1`
+
 ## [1.7.7](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-1.7.6...@stream-io/video-react-bindings-1.7.7) (2025-07-11)
 
 ### Dependency Updates