+++ packages/react-sdk/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.18.13](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.18.12...@stream-io/video-react-sdk-1.18.13) (2025-07-17)
+
+### Dependency Updates
+
+- `@stream-io/video-client` updated to version `1.26.1`
+- `@stream-io/video-react-bindings` updated to version `1.7.8`
+
 ## [1.18.12](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.18.11...@stream-io/video-react-sdk-1.18.12) (2025-07-11)
 
 ### Dependency Updates